@puneeshkhanna 2024-01-16
@regisss  - this PR should provide the same gains as we saw in https://github.com/huggingface/optimum-habana/pull/626/ for TP-4 and TP-2 cases of llama70B inference. 
I just need to test once more with the final changes that I have pushed and will update the results in comments.