@puneeshkhanna 2024-01-16
@regisss - Also can we label this too with synapse 1.14 since it is dependent on https://github.com/huggingface/optimum-habana/pull/626/