@puneeshkhanna 2024-01-17
yes @regisss - we should first merge https://github.com/huggingface/optimum-habana/pull/626 and then this one to avoid merge conflicts hopefully.