@puneeshkhanna 2024-01-17
@schoi-habana  - Will you get a chance to verify finetuning too with the changes in this PR. Note that these changes are application when flash attention is disabled.